Add back vim-sleuth
It helps when opening new lines in a e.g. Lua or TypeScript file.
Formatting should use spaces for this to work better. I tried using the
Biome's default with tabs but could get Neovim to work nicely with both
.ts and .lua files then (since stylua uses spaces, probably). I
therefore now only use spaces with Biome (and other formatters).

!fixup